legalnova_chatbot/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
google-generativeai==0.8.3
requests==2.32.3
pymupdf==1.24.10
python-docx==1.1.2
gunicorn==23.0.0